+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# v1.0.4:
+- Fixed a bug when several matches are happening at the same time
+- Customized discord.ext.tasks as lib.tasks to have better flexibility on tasks
+
+# v1.0.3:
+- It is no longer possible to register with a character that is already registered
+- Team captains can now select a map
+- Added =confirm command for Team Captains to agree on a map
+- Added role updates when agreeing with the rules
+- Added notify feature
+- Added =unregister @player to remove a player from the system (including db)
+
 # v1.0.2:
 - Now properly checking if player have no missing faction when registering with a Jaeger char
 - Lobby size can now be modified from the config file